requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pyttsx3==2.90
SpeechRecognition==3.10.0
cohere==4.37
torch==2.0.1
ultralytics==8.0.196
easyocr==1.7.0
twilio==8.10.0
python-dotenv==1.0.0
PyAudio==0.2.11 